Renamed original class and fix import
